app/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
pydantic==2.7.1
pydantic-settings==2.2.1
python-dotenv==1.0.1
asyncpg==0.29.0
sqlalchemy==2.0.30
passlib[bcrypt]==1.7.4
bcrypt==3.2.2
python-jose[cryptography]==3.3.0
httpx==0.27.0
authlib==1.3.0
aiosmtplib==2.0.0
alembic==1.12.0
itsdangerous>=2.0
psycopg2-binary==2.9.9
openai==1.26.0
nltk
python-multipart
